🔧 fix: Modifying to a Server Component
Ref #51
